requirements.txt
scrapling[all]
gspread
google-auth